refactor: S3ObjectClient를 infrastructure 패키지로 이동
