+++ backend/src/test/java/com/staccato/TestConfig.java
@@ -3,7 +3,8 @@ package com.staccato;
 import org.springframework.boot.test.context.TestConfiguration;
 import org.springframework.context.annotation.Bean;
 
-import com.staccato.image.domain.S3ObjectClient;
+import com.staccato.infrastructure.FakeS3ObjectClient;
+import com.staccato.infrastructure.S3ObjectClient;
 
 @TestConfiguration
 public class TestConfig {
